+++ src/components/InputSelect/types.ts
@@ -1,0 +1,8 @@
+export interface InputSelectData {
+  id: string;
+  inFocus: boolean;
+  _uid: string;
+  selected: string | Record<string, unknown>;
+}
+
+export type InputValueType = string | number;