+++ CHANGELOG.md
@@ -4,11 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
 ## [Unreleased]
 
-## [0.1.636] - 2022-01-20
+## [0.1.639] - 2022-02-15
 ### Add
 - Husky git precommit
+- InputPhone label style
+
+## [0.1.638] - 2022-01-27
+### Fix
+- Add simple flag to Autocomplete for back competibility
+
+## [0.1.637] - 2022-01-27
+## [0.1.636] - 2022-01-24
+### Fix
+- Add selectedObj to Autocomplete
 
 ## [0.1.635] - 2022-01-13
 ### Fix